Move uses_api_lock alongside the other decorators
